A simple change that adds the option to enable socket reuse when calling listen. This is necessary to get clustering to work properly.