@alexhultman 2018-03-10
HTTP is not even stable, not going to add new "perf options" prematurely