@Dmitry-N-Medvedev 2018-03-10
how does socket reuse flag relate to HTTP in the first place?